sixth version

+++ app.py
@@ -1,4 +1,4 @@
-num =int(input("enter a number"))
+num =float(input("enter a number"))
 if num % 2 == 0:
      print("even number")
 else:
